* Added VC2017 project for RTConsole, deleted MSVC2005 one * Big change with RTPack - tools directory is gone and RTPack is now a normal project * RTPack no longer requires CLanlib V1.0 to build, all functions switched over to use Proton classes like SoftSurface * RTPack now optionally builds without PVRTC support (if RT_NO_PVR is defined.  The VC2017 project has an added configuration setup for this mode) * Linux media building scripts moved to /RTPack/linux but still function mostly the same (Aki set that up a while ago, I did some small tweaks so it still works) * JPGSurfaceLoader class can now also save JPG files out (yeah, great file name..) * All sample projects build on Raspberry Pi (RTSimpleApp/linux/linux_compile.sh etc)

+++ shared/util/MiscUtils.cpp
@@ -291,7 +291,7 @@ string DataToByteHexDisplay(string data, int maxNumCharsToShow)
 	string hexData = "Hex data: ";
 
 	char temp[255];
-	for (int i = 0; i < maxNumCharsToShow && i < data.length(); i++)
+	for (int i = 0; i < maxNumCharsToShow && i < (int)data.length(); i++)
 	{
 		sprintf(temp, "%hhx,", data[i]);
 		hexData += (string(temp) + " ");
